Finaliza template de e-mail padrão
