--- sme_terceirizadas/dados_comuns/static/css/email_base.css
@@ -1,47 +1,0 @@
-.conteiner-tex {
-
-}
-
-.email-cabecalho {
-  padding: 1em 1em 0.6em 1em;
-  background: #FFFFFF;
-}
-
-.bg-email {
-  background: #F5F6F8;
-  height: 500px;
-  top: 0;
-}
-
-.imagem-email {
-  position: absolute;
-}
-
-.email-conteiner {
-  position: absolute;
-  align-content: center;
-  width: 610px;
-  background: #888888;
-}
-
-.retangulo-conteudo {
-
-  border-radius: 4px;
-  text-align: center;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-  color: #42474A;
-  background-color: #FFFFFF;
-  padding: 5em;
-  margin-left: 3em;
-  margin-right: 3em;
-  font-size: 14px;
-  position: absolute;
-  z-index: 50;
-  margin-top: 6em;
-}
-
-.email-footer {
-  padding: 2em 1em 0.5em 1em;
-  color: #198459;
-  background: #F5F6F8;
-}
