--- sme_terceirizadas/dados_comuns/templates/test.html
@@ -1,47 +1,0 @@
-"""
-<html lang="pt-br">
-<head>
-  <meta charset="UTF-8">
-  <title>SIGPAE - Notificação</title>
-
-  <!-- Bootstrap core CSS -->
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
-        integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-
-
-  <link rel="stylesheet" href="http://localhost:8000/media/email/static/css/email_base.css">
-  <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet">
-</head>
-<body class="email-conteiner conteiner-tex">
-<div>
-  <div class="email-cabecalho">
-
-    <img src="http://localhost:8000/media/email/static/images/logo-sigpae.svg" alt=""/>
-  </div>
-  <div class="bg-email">
-    <img class="imagem-email" src="http://localhost:8000/media/email/static/images/bg-email.svg" alt=""/>
-    <div class="retangulo-conteudo">
-      <p>Conteúdo de texto do e-mail - Lorem ipsum dolor sit amet, consectetur adipiscing elit. Iaculis sagittis ut
-        ullamcorper natoque praesent duis vulputate id magna. </p>
-      <p>O conteúdo que se encontra nesse layout é somente um Exemplo de Conteúdo. Lorem ut ullamcorper natoque praesent
-        duis vulputate id magna. </p>
-    </div>
-  </div>
-
-</div>
-</body>
-<footer class="email-footer">
-  <div class="row">
-    <div class="col">
-      <img src="http://localhost:8000/media/email/static/images/logo-sp-color.svg" alt=""/>
-    </div>
-    <div class="col">
-      <span>contatodeemail@sigpae.sp.gov.br</span>
-    </div>
-    <div class="col">
-      0800-123 4567
-    </div>
-  </div>
-</footer>
-</html>
-"""
